move all main components into folder components/product/main
move all part components into folder components/product/parts
