+++ frontend/src/components/product/main/CriticalProductList.css
