+++ frontend/src/components/product/main/ProductChangeLog.tsx
@@ -1,7 +1,7 @@
 import './ProductChangeLog.css';
 import React from 'react';
 import {ProductChange} from "../../../types/ProductChange.ts";
-import ProductChangeCard from "../../parts/ProductChangeCard.tsx";
+import ProductChangeCard from "../parts/ProductChangeCard.tsx";
 import TableHead from "../../htmlParts/TableHead.tsx";
 
 type ChangeLogProps = {
