+++ frontend/src/components/product/main/ProductDetail.css
