+++ frontend/src/components/product/parts/Barcode.tsx
