+++ frontend/src/components/product/parts/ProductDtoCard.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {ProductDTO} from '../../types/ProductDTO.ts';
+import {ProductDTO} from '../../../types/ProductDTO.ts';
 import './ProductDtoCard.css';
 
 type ProductCardProps = {
